Extract read count tooltip formatter to util
